solve typical90-p

+++ atcoder/other/typical90/typical90-p.cpp
@@ -5,12 +5,12 @@ using ll = long long;
 #define se second
 #define all(a) a.begin(),a.end()
 int main(){
-  int n,a,b,c;
+  ll n,a,b,c;
   cin >> n >> a >> b >> c;
-  int ans=10000;
-  for(int i=0; i<10000; ++i){
-    for(int j=0; j<10000; ++j){
-      int bar=n-a*i-b*j;
+  ll ans=10000;
+  for(ll i=0; i<10000; ++i){
+    for(ll j=0; j<10000; ++j){
+      ll bar=n-a*i-b*j;
       if(bar>=0 && bar%c==0) ans=min(i+j+bar/c,ans);
     }
   }
